pyproject.toml
[project]
name = "pdf-to-markdown"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "markdown2>=2.5.3",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
    "pytesseract>=0.3.13",
    "streamlit>=1.46.1",
]
